add new

+++ app/Http/Controllers/PostController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Post;
 use App\Tag;
+use App\PostTag;
+use App\User;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Response;
 use RealRashid\SweetAlert\Facades\Alert;
@@ -12,8 +14,9 @@ use RealRashid\SweetAlert\Facades\Alert;
 class PostController extends Controller
 {
 
-    public function __construct(){
-        $this -> middleware('auth')->only(['index', 'create', 'store', 'edit', 'update', 'destroy']);
+    public function __construct()
+    {
+        $this->middleware('auth')->only(['index', 'create', 'store', 'edit', 'update', 'destroy']);
     }
 
     /**
@@ -24,8 +27,10 @@ class PostController extends Controller
     public function index()
     {
         $posts = Post::all();
+        $tags = Tag::limit(3)->get();
+        $users = User::inRandomOrder()->limit(3)->get()->except(Auth::id());
 
-        return view('post.index', compact('posts'));
+        return view('post.index', compact('posts', 'tags', 'users'));
     }
 
     /**
@@ -45,39 +50,38 @@ class PostController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
-    {        
+    {
         $post = new Post;
         $user = Auth::user();
-        
-        $post->content = $request['content']; 
-        $post->user_id = $user->id; 
+
+        $post->caption = $request['caption'];
+        $post->user_id = $user->id;
 
         $tags_arr = explode(',', $request['tags']);
         $tag_ids = [];
-        foreach($tags_arr as $tag_name){
+        foreach ($tags_arr as $tag_name) {
             $tag = Tag::firstOrCreate([
-                   'tag_name' => $tag_name
+                'tag_name' => $tag_name
             ]);
             $tag_ids[] = $tag->id;
         }
 
-        if ($request -> hasFile('image')){
-            $file = $request -> file('image');
-            $extension = $file -> getClientOriginalExtension();
+        if ($request->hasFile('img_url')) {
+            $file = $request->file('img_url');
+            $extension = $file->getClientOriginalExtension();
             $filename = time() . '.' . $extension;
-            $file -> move('uploads/post', $filename);
-            $post -> image = $filename;
-        }
-        else{
+            $file->move('uploads/post', $filename);
+            $post->img_url = $filename;
+        } else {
             /* return $request; */
-            $post -> image = '';
+            $post->img_url = '';
         }
-        
+
         $post->save();
         $post->tags()->sync($tag_ids);
 
-        Alert::success('Success', 'Tweet added');
-        
+        Alert::success('Success', 'Post added');
+
         return redirect('/post');
     }
 
@@ -103,8 +107,20 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if ($post->user_id != Auth::id()) {
+            return redirect('post');
+        }
 
-        return view('post.edit', compact('post'));
+        $post_tags = PostTag::where('post_id', '=', $id)->get();
+        $name_arr = [];
+        foreach ($post_tags as $tags) {
+            $temp_tags = Tag::find($tags->tag_id);
+            $name_arr[] = $temp_tags->tag_name;
+        }
+
+        $tag_name = implode(',', $name_arr);
+
+        return view('post.edit', compact('post', 'tag_name'));
     }
 
     /**
@@ -116,10 +132,25 @@ class PostController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $post = Post::find($id);
+        if ($post->user_id != Auth::id()) {
+            return redirect('post');
+        }
+
         $update = Post::where('id', $id)
-                  ->update([
-                    "content" => $request["content"]
-                  ]);
+            ->update([
+                "caption" => $request["caption"]
+            ]);
+
+        $tags_arr = explode(',', $request['tags']);
+        $tag_ids = [];
+        foreach ($tags_arr as $tag_name) {
+            $tag = Tag::firstOrCreate([
+                'tag_name' => $tag_name
+            ]);
+            $tag_ids[] = $tag->id;
+        }
+        $post->tags()->sync($tag_ids);
 
         return redirect('post');
     }
@@ -132,7 +163,12 @@ class PostController extends Controller
      */
     public function destroy($id)
     {
-        Post::destroy($id);   
+        $post = Post::find($id);
+        if ($post->user_id != Auth::id()) {
+            return redirect('post');
+        }
+
+        Post::destroy($id);
 
         return redirect('post');
     }
